Administrative hierarchy

Retrieve polygon geometries of municipalities in Germany with details on their relationships to administrative areas of higher levels in the territorial hierarchy. The output of this functions contains the identifiers and names of the NUTS1 to NUTS3 areas that each municipality belongs to.

Usage

bkg_admin_hierarchy(
  key_date = c("0101", "1231"),
  year = "latest",
  timeout = 120,
  update_cache = FALSE
)

Arguments

key_date

For resolution %in% c("250", "5000"), specifies the key date from which to download administrative data. Can be either "0101" (January 1) or "1231" (December 31). The latter is able to georeference statistical data while the first integrates changes made in the new year. If "1231", population data is attached, otherwise not. Note that population data is not available at all scales (usually 250 and 1000). Defaults to "0101".

year

Version year of the dataset. You can use latest to retrieve the latest dataset version available on the BKG's geodata center. Older versions can be browsed using the archive.

timeout

Timeout value for the data download passed to req_timeout. Adjust this if your internet connection is slow or you are downloading larger datasets.

update_cache

By default, downloaded files are cached in the tempdir() directory of R. When downloading the same data again, the data is not downloaded but instead taken from the cache. Sometimes this can be not the desired behavior. If you want to overwrite the cache, pass TRUE. Defaults to FALSE, i.e. always adopt the cache if possible.

Value

An sf tibble with multipolygon geometries similar to the output of bkg_admin(level = "gem"). The tibble additionally contains columns NUTS*_CODE and NUTS*_NAME giving the identifiers and names of the administrative areas the municipalities belong to.

Examples

bkg_admin_hierarchy()
#> Error in httr2::req_perform(req, path = path): Failed to perform HTTP request.
#> Caused by error in `curl::curl_fetch_disk()`:
#> ! Timeout was reached [daten.gdz.bkg.bund.de]:
#> Operation timed out after 120000 milliseconds with 21304468 out of 24049413 bytes received